Align contract versions (#1213)
* Align contract versions

* fix tests

+++ packages/coins/src/interfaces/IDeployedCoinVersionLookup.sol
@@ -7,5 +7,5 @@ interface IDeployedCoinVersionLookup {
     /// @notice Gets the version for a deployed coin
     /// @param coin The address of the coin
     /// @return version The version of the coin (0 if not found)
-    function getVersionForDeployedCoin(address coin) external view returns (uint32);
+    function getVersionForDeployedCoin(address coin) external view returns (uint8);
 }
